+++ src/short-urls/helpers/ShortUrlsRow.js
@@ -6,11 +6,14 @@ import { shortUrlsListParamsType } from '../reducers/shortUrlsListParams';
 import { serverType } from '../../servers/prop-types';
 import ExternalLink from '../../utils/ExternalLink';
 import { shortUrlType } from '../reducers/shortUrlsList';
-import { stateFlagTimeout } from '../../utils/utils';
 import Tag from '../../tags/helpers/Tag';
 import './ShortUrlsRow.scss';
 
-const ShortUrlsRow = (ShortUrlsRowMenu, colorGenerator) => class ShortUrlsRow extends React.Component {
+const ShortUrlsRow = (
+  ShortUrlsRowMenu,
+  colorGenerator,
+  stateFlagTimeout
+) => class ShortUrlsRow extends React.Component {
   static propTypes = {
     refreshList: PropTypes.func,
     shortUrlsListParams: shortUrlsListParamsType,