+++ src/utils/utils.js
@@ -1,6 +1,16 @@
+import L from 'leaflet';
+import marker2x from 'leaflet/dist/images/marker-icon-2x.png';
+import marker from 'leaflet/dist/images/marker-icon.png';
+import markerShadow from 'leaflet/dist/images/marker-shadow.png';
+
 const DEFAULT_TIMEOUT_DELAY = 2000;
 
-export const stateFlagTimeout = (setState, flagName, initialValue = true, delay = DEFAULT_TIMEOUT_DELAY) => {
+export const stateFlagTimeout = (setTimeout) => (
+  setState,
+  flagName,
+  initialValue = true,
+  delay = DEFAULT_TIMEOUT_DELAY
+) => {
   setState({ [flagName]: initialValue });
   setTimeout(() => setState({ [flagName]: !initialValue }), delay);
 };
@@ -16,4 +26,14 @@ export const determineOrderDir = (clickedField, currentOrderField, currentOrderD
   };
 
   return currentOrderDir ? newOrderMap[currentOrderDir] : 'ASC';
+};
+
+export const fixLeafletIcons = () => {
+  delete L.Icon.Default.prototype._getIconUrl;
+
+  L.Icon.Default.mergeOptions({
+    iconRetinaUrl: marker2x,
+    iconUrl: marker,
+    shadowUrl: markerShadow,
+  });
 };