+++ src/visits/SortableBarGraph.js
@@ -1,14 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { fromPairs, head, identity, keys, pipe, prop, reverse, sortBy, toLower, toPairs, type } from 'ramda';
+import { fromPairs, head, keys, pipe, prop, reverse, sortBy, toLower, toPairs, type } from 'ramda';
 import SortingDropdown from '../utils/SortingDropdown';
 import GraphCard from './GraphCard';
+
+const toLowerIfString = (value) => type(value) === 'String' ? toLower(value) : value;
 
 export default class SortableBarGraph extends React.Component {
   static propTypes = {
     stats: PropTypes.object.isRequired,
     title: PropTypes.string.isRequired,
     sortingItems: PropTypes.object.isRequired,
+    extraHeaderContent: PropTypes.arrayOf(PropTypes.func),
   };
 
   state = {
@@ -17,13 +20,12 @@ export default class SortableBarGraph extends React.Component {
   };
 
   render() {
-    const { stats, sortingItems, title } = this.props;
+    const { stats, sortingItems, title, extraHeaderContent } = this.props;
     const sortStats = () => {
       if (!this.state.orderField) {
         return stats;
       }
 
-      const toLowerIfString = (value) => type(value) === 'String' ? toLower(value) : identity(value);
       const sortedPairs = sortBy(
         pipe(
           prop(this.state.orderField === head(keys(sortingItems)) ? 0 : 1),
@@ -48,6 +50,11 @@ export default class SortableBarGraph extends React.Component {
             onChange={(orderField, orderDir) => this.setState({ orderField, orderDir })}
           />
         </div>
+        {extraHeaderContent && extraHeaderContent.map((content, index) => (
+          <div key={index} className="float-right">
+            {content()}
+          </div>
+        ))}
       </GraphCard>
     );
   }